This is a c++ base code that I put together.
It does not follow the sharkattack pokemon specification that was written up by winxordie.
I am not implying that we do the project in c++, but rather just putting this online rather than having it sit on my computer .
It is just a window where you can control a character, I believe the project is well OOP-ed out so it can be easily reused to form-fit our needs if we decide so.

-Requires SDL,SDL_image and Boost Libraries.

I have tested it on CISE machines with eclipse CDT plugin (they have updated to new eclipse version 3.5 which made this possible), as well as my own laptop which uses Ubuntu... I doubt it will work in windows since it likely requires extra code to create a 'Windows' window, but have a try.

Revision 1?


git-svn-id: http://prmf.googlecode.com/svn@307 93ad656a-1fb5-1e14-4869-072bfc7f1330