Sprite Splicer
This is a work in progress which is likely to never be completed though.

It is going to be an implementation of libpng which is able to
select sprites from a spritesheet and return their locations.

Useful sources:
http://www.libpng.org/pub/png/libpng-1.4.0-manual.pdf
http://www.piko3d.com/?page_id=68

The printf statements are only there for debugging purpose, I want the actual
end product to just take in a filename and return an array of box structs indicating
the locations and sizes of each sprite on the sheet.

git-svn-id: http://prmf.googlecode.com/svn@308 93ad656a-1fb5-1e14-4869-072bfc7f1330