Fixed some surface bugs with !firstsolved and !datesolved. Beyond the checking of problem IDs to make sure that they are of the valid format (alphanumeric and _, between 3-8 characters), SPOJBot cannot currently distinguish between invalid problem codes and valid ones. This would involve additional checking that would significantly slow down the !sl/!spojlink command.
Also migrated some code to Utility for reuse and replaced some duplicates, triplicates, and further multiplicates of code with function calls.

Oh, and the users txt files in this update are up to date as of 10-2-11.

Cry of bewilderment: The more I work on this bot, the more confused I get. Why is there code duplicated or triplicated among SPOJBot, User, and Utility? What's the purpose of Utility when utility methods can be found in all three classes? Should I even bother fixing it, or should I just maintain the convention and duplicate my own code...?

git-svn-id: http://prmf.googlecode.com/svn@539 93ad656a-1fb5-1e14-4869-072bfc7f1330